requirements.txt
# Document processing
python-docx==1.1.2

# Embeddings & Vector Store
sentence-transformers>=2.7.0,<3.3
faiss-cpu>=1.7.0
numpy<2.0

# LLM (gọi Gemini qua REST API)
requests>=2.28.0

# Web server
flask>=2.3.0

# Environment variables
python-dotenv>=1.0.0
